Process RestoreBodyPart in Packet rather than Patch
